requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
openai==1.3.7
matplotlib==3.8.2
python-multipart==0.0.6
python-dotenv==1.1.0
sqlalchemy==2.0.41
ruff==0.11.13
mypy==1.16.1